backend/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
gunicorn
psycopg2-binary
whitenoise
python-dotenv

